feat(entities-plugins): rename ai-mcp to ai-mcp-proxy
