--- packages/entities/entities-plugins/src/components/free-form/AIMcp/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AIMcpForm.vue'
