+++ packages/entities/entities-plugins/src/components/free-form/AIMcpProxy/AIMcpProxyForm.vue
