+++ packages/entities/entities-plugins/src/components/free-form/AIMcpProxy/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AIMcpProxyForm.vue'
