+++ packages/entities/entities-plugins/src/components/free-form/AIMcpProxy/utils.ts
