+++ packages/entities/entities-plugins/src/components/free-form/index.ts
@@ -1,5 +1,5 @@
 export { default as RequestCalloutForm } from './RequestCallout'
 export { default as ServiceProtectionForm } from './ServiceProtection'
 export { default as DatakitForm } from './Datakit'
-export { default as AIMcpForm } from './AIMcp'
+export { default as AIMcpProxyForm } from './AIMcpProxy'
 export { default as CommonForm } from './Common'
